# Reduce trivial warning messages from `tests/sampler_tests`