<!-- Thank you for creating a pull request! In general, we merge your pull request after it gets two or more approvals. To proceed to the review process by the maintainers, please make sure that the PR meets the following conditions: (1) it passes all CI checks, and (2) it is neither in the draft nor WIP state. If you wish to discuss the PR in the draft state or need any other help, please mention the Optuna development team in the PR. -->

## Motivation
<!-- Describe your motivation why you will submit this PR. This is useful for reviewers to understand the context of PR. -->

A part of #3815. 

## Description of the changes
<!-- Describe the changes in this PR. -->

- Resolve `step` related warning message such as `suggest_int(low=1, high=10, step=2`)
- Ignore the expected `UserWarning`
- Resolve numpy warning message via TPE (see the following my comment too)


Note that there are still several warning messages from TPE sampler and BoTorchSampler, I don't remove such messages because I suppose such changes make this PR more complicated.