@contramundum53 2022-08-25
I ran the speed benchmark on MacBookAir(M1).

Master:
```
              ----------------------------- ------------
                  inmemory, random, 1000      193±8ms
                 inmemory, random, 10000     4.54±0.1s
                   inmemory, tpe, 1000       3.75±0.06s
                  inmemory, cmaes, 1000       548±8ms
                   sqlite, random, 1000      9.82±0.09s
               cached_sqlite, random, 1000   9.85±0.03s
                   redis, random, 1000       6.58±0.03s
                cached_redis, random, 1000   6.55±0.01s
              ============================= ============
```

This PR:
```
              ============================= ============
                storage, sampler, n_trials
              ----------------------------- ------------
                  inmemory, random, 1000      186±3ms
                 inmemory, random, 10000     4.43±0.2s
                   inmemory, tpe, 1000       3.76±0.03s
                  inmemory, cmaes, 1000       525±6ms
                   sqlite, random, 1000      11.0±0.6s
               cached_sqlite, random, 1000   10.4±0.4s
                   redis, random, 1000       6.58±0.02s
                cached_redis, random, 1000   6.52±0.02s
              ============================= ============
```
There is little performance regression. LGTM.